+++ CHANGELOG.md
@@ -22,6 +22,9 @@ speaking, considered to be the player's experience.
 
 ### Changed
 
+- Changed the graphics of Standard tileset terrain that looked like it should
+  block building and movement but didn't, and vice versa, so that it's now
+  obvious by just looking at it.
 - Changed default keybind for train soldier to `F` from `W`, to make it more
   consistent with the other fort-related keybinds (build fort).